+++ R/globals.R
@@ -2,4 +2,4 @@
 
 # Global bindings for package checks
 utils::globalVariables(c("variable", "column", "study", "home", "room", "config",
-                         "ID"))
+                         "ID", "N", "NAs"))